crewup/requirements.txt
Django==6.0.3
djangorestframework==3.17.1
django-filter==25.2
python-decouple==3.8
whitenoise==6.12.0
Pillow==12.1.1
